deployment_package/requirements.txt
# This file can contain package requirements for the deployment to be installed via Python pip. Packages defined here
# will be installed with your deployment automatically.
#
# Example:
# requests==2.24.0
pandas==1.2.1
pysharepoint==0.1.7
python-dotenv==0.17.1